+++ src/index.js
@@ -2,11 +2,13 @@ import contentType from './contentType.js'
 import parse from './parse.js'
 import extract from './extract.js'
 import request, { createRequest } from './request.js'
+import RedirectRefusedError from './RedirectRefusedError.js'
 
 export default {
   contentType,
   parse,
   extract,
   request,
-  createRequest
+  createRequest,
+  RedirectRefusedError
 }